+++ internal/models/userModel.go
@@ -1,10 +1,11 @@
 package model
 
 import (
-	"context"
-	"fmt"
 	"go-server/internal/repositories/db/postgresUser"
 	"go-server/pkg/logging"
+
+	"context"
+	"fmt"
 
 	"github.com/google/uuid"
 )
@@ -16,15 +17,16 @@ type User struct {
 }
 
 func (usr *User) Save() (interface{}, error) {
-	var data db.UserData
-	data.Username = usr.Username
-	data.Email = usr.Email
-
 	logger := logging.GetLogger()
 	repo := db.NewRepository(logger)
+
 	if repo == nil {
 		return nil, fmt.Errorf("failed to create repository")
 	}
+
+	var data db.UserData
+	data.Username = usr.Username
+	data.Email = usr.Email
 
 	if usr.UserId != uuid.Nil {
 		data.UserId = usr.UserId
@@ -34,15 +36,21 @@ func (usr *User) Save() (interface{}, error) {
 	}
 }
 
-func NewUser(Username, Email string) *User {
-	usr := new(User)
-	usr.Username = Username
-	usr.Email = Email
-	return usr
+func NewUser(username, email string) *User {
+	return &User{
+		Username: username,
+		Email:    email,
+	}
 }
+
 func LoadUser(id string) (*User, error) {
 	logger := logging.GetLogger()
 	repo := db.NewRepository(logger)
+
+	if repo == nil {
+		return nil, fmt.Errorf("failed to create repository")
+	}
+
 	data, err := repo.FindOne(context.TODO(), id)
 	if err != nil {
 		logger.Infof("Failed to load User: %v", err)
@@ -59,6 +67,11 @@ func LoadUser(id string) (*User, error) {
 func LoadUsers() ([]*User, error) {
 	logger := logging.GetLogger()
 	repo := db.NewRepository(logger)
+
+	if repo == nil {
+		return nil, fmt.Errorf("failed to create repository")
+	}
+
 	data, err := repo.FindAll(context.TODO())
 	if err != nil {
 		logger.Infof("Failed to load Users: %v", err)
@@ -76,9 +89,15 @@ func LoadUsers() ([]*User, error) {
 	return usrs, nil
 
 }
+
 func LoadUserByEmail(email string) (*User, error) {
 	logger := logging.GetLogger()
 	repo := db.NewRepository(logger)
+
+	if repo == nil {
+		return nil, fmt.Errorf("failed to create repository")
+	}
+
 	data, err := repo.FindUserByEmail(context.TODO(), email)
 	if err != nil {
 		logger.Infof("Failed to load User by email: %v", err)
@@ -94,6 +113,11 @@ func LoadUserByEmail(email string) (*User, error) {
 func DeleteUser(id string) error {
 	logger := logging.GetLogger()
 	repo := db.NewRepository(logger)
+
+	if repo == nil {
+		return fmt.Errorf("failed to create repository")
+	}
+
 	if err := repo.Delete(context.TODO(), id); err != nil {
 		logger.Infof("Failed to delete User: %v", err)
 		return err
